requirements.txt
# ── Web framework ──────────────────────────────
flask>=3.0.0
werkzeug>=3.0.0
pyngrok>=8.0.0

# ── GPU monitoring ─────────────────────────────
pynvml>=13.0.0

# ── AI / Training ──────────────────────────────
# torch + unsloth: install separately via pip based on CUDA version
# pip install torch==2.8.0 --index-url https://download.pytorch.org/whl/cu121
# pip install "unsloth[cu121-torch280] @ git+https://github.com/unslothai/unsloth"
trl==0.22.2
transformers==5.2.0
datasets>=4.0.0
huggingface-hub>=0.25.0
peft>=0.14.0
bitsandbytes>=0.45.0
accelerate>=1.0.0
sentence-transformers>=3.0.0

# ── Evaluation ─────────────────────────────────
anthropic>=0.89.0
nltk>=3.9.0
rouge-score>=0.1.2
scikit-learn>=1.6.0

# ── RAG / Web search ───────────────────────────
duckduckgo-search>=8.0.0

# ── Utilities ──────────────────────────────────
numpy>=2.0.0
python-dotenv>=1.0.0
